// src/app/models/Bases/basesImpugnar.ts
export class BasesImpugnar{
    temp_DATE: number;
    temp_UNDERLYINGCLASIFICATION: number;
    temp_DESC_UNDERLYINGCLASIFICATION: string;
    temp_MonthContract: string;
    temp_ASK: number;
    temp_BID: number;
    temp_MID: number;
    temp_Impugnar:Boolean;
    temp_OrigenImpugnacion: string;
    temp_Valor:number;
    temp_FechaTermino:Date;
    temp_UltValor: number;
    temp_UltFechaTermino: Date;
    temp_SustentosAsociados: String;
    temp_ID_BaseImpugn: number;
    temp_ID_MonthContract: number;
    temp_ApprovalStatus: string;
}