// src/app/models/Bases/impugnacion.ts
export class Impugnacion{
    t439_ID: number;
    t439_Date: number;
    t439_UnderlyingClassification: number;
    t439_MonthContract: number;
    t439_EffectiveDate: number;
    t439_Before: number;
    t439_Now: number;
    t439_Hour: string;
    t439_User: string;
    t439_Status: boolean;
    t439_OriginImpugn: number;
}